Move typeutil to generic util package
